FEATURE: sample-app-lose-weight-编译运行

+++ sample-app-lose-weight/src/main/java/com/allever/lose/weight/ui/ReminderFragment.java
@@ -4,6 +4,8 @@ import android.app.TimePickerDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
 import androidx.annotation.Nullable;
+
+import com.chad.library.adapter.base.listener.OnItemChildClickListener;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -29,24 +31,16 @@ import com.allever.lose.weight.ui.dialog.WeekChoiceDialogFragment;
 
 import java.util.List;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import butterknife.Unbinder;
-
 
 /**
  * Created by Mac on 2018/3/1.
  */
 
 public class ReminderFragment extends BaseFragment<IReminderView, ReminderPresenter> implements IReminderView {
-    @BindView(R.id.toolbar)
     Toolbar mToolbar;
-    @BindView(R.id.recycler_view)
     RecyclerView mRecyclerView;
-    Unbinder unbinder;
-    @BindView(R.id.add_reminder)
     FloatingActionButton addReminder;
+
     public static final String TAG = "ReminderFragment";
     private RemindRecAdapter mAdapter;
 
@@ -63,7 +57,36 @@ public class ReminderFragment extends BaseFragment<IReminderView, ReminderPresen
 
         View view = inflater.inflate(R.layout.fragment_reminder, container, false);
 
-        unbinder = ButterKnife.bind(this, view);
+        mToolbar = view.findViewById(R.id.toolbar);
+        mRecyclerView = view.findViewById(R.id.recycler_view);
+        addReminder = view.findViewById(R.id.add_reminder);
+        addReminder.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                final Config.Reminder reminder = new Config.Reminder();
+                TimePickerFragment timePickerFragment = new TimePickerFragment();
+                timePickerFragment.show(getFragmentManager(), new TimePickerDialog.OnTimeSetListener() {
+                    @Override
+                    public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
+                        Log.i(TAG, "hour:" + hourOfDay + "min:" + minute);
+                        reminder.setHour(hourOfDay);
+                        reminder.setMinute(minute);
+                        reminder.setSecond(0);
+                        reminder.setRemindSwitch(true);
+                        reminder.setSunRepeat(true);
+                        reminder.setMonRepeat(true);
+                        reminder.setTueRepeat(true);
+                        reminder.setWebRepeat(true);
+                        reminder.setThurRepeat(true);
+                        reminder.setFriRepeat(true);
+                        reminder.setSatRepeat(true);
+
+                        showWeek(reminder);
+                    }
+                });
+            }
+        });
+
 
         initToolbar(mToolbar, R.string.remind);
 
@@ -78,73 +101,35 @@ public class ReminderFragment extends BaseFragment<IReminderView, ReminderPresen
         mRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         mRecyclerView.setHasFixedSize(true);
         mRecyclerView.addItemDecoration(new SpacesItemDecoration(ScreenUtils.dp2px(10)));
-        mAdapter.setOnItemChildClickListener(new BaseQuickAdapter.OnItemChildClickListener() {
+        mAdapter.setOnItemChildClickListener(new OnItemChildClickListener() {
             @Override
             public void onItemChildClick(BaseQuickAdapter adapter, View view, int position) {
                 Log.d(TAG, "onItemChildClick: reminderItem position = " + position);
-                switch (view.getId()){
-                    case R.id.delete:
-                        //删除提醒记录
-                        mPresenter.deleteReminder(mReminderList.get(position));
-                        break;
-                    case R.id.reminder_switch:
-                        Switch aSwitch = (Switch) view.findViewById(R.id.reminder_switch);
-                        Config.Reminder reminder = mReminderList.get(position);
-                        reminder.setRemindSwitch(aSwitch.isChecked());
-                        if (aSwitch.isChecked()){
-                            mPresenter.setReminder(_mActivity,mReminderList.get(position));
-                        }else {
-                            mPresenter.cancelReminder(_mActivity, mReminderList.get(position));
-                        }
-                        //更新数据库
-                        mPresenter.updateReminder(position, reminder);
-                        break;
-                    case R.id.id_item_remind_tv_repeat:
-                        showWeek(mReminderList.get(position));
-                        break;
-                    default:
-                        break;
+                int id = view.getId();
+                if (id == R.id.delete) {//删除提醒记录
+                    mPresenter.deleteReminder(mReminderList.get(position));
+                } else if (id == R.id.reminder_switch) {
+                    Switch aSwitch = (Switch) view.findViewById(R.id.reminder_switch);
+                    Config.Reminder reminder = mReminderList.get(position);
+                    reminder.setRemindSwitch(aSwitch.isChecked());
+                    if (aSwitch.isChecked()) {
+                        mPresenter.setReminder(_mActivity, mReminderList.get(position));
+                    } else {
+                        mPresenter.cancelReminder(_mActivity, mReminderList.get(position));
+                    }
+                    //更新数据库
+                    mPresenter.updateReminder(position, reminder);
+                } else if (id == R.id.id_item_remind_tv_repeat) {
+                    showWeek(mReminderList.get(position));
                 }
             }
         });
         mRecyclerView.setAdapter(mAdapter);
     }
 
-
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        unbinder.unbind();
-    }
-
     @Override
     protected ReminderPresenter createPresenter() {
         return new ReminderPresenter();
-    }
-
-    @OnClick(R.id.add_reminder)
-    public void onViewClicked() {
-        final Config.Reminder reminder = new Config.Reminder();
-        TimePickerFragment timePickerFragment = new TimePickerFragment();
-        timePickerFragment.show(getFragmentManager(), new TimePickerDialog.OnTimeSetListener() {
-            @Override
-            public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
-                Log.i(TAG, "hour:" + hourOfDay + "min:" + minute);
-                reminder.setHour(hourOfDay);
-                reminder.setMinute(minute);
-                reminder.setSecond(0);
-                reminder.setRemindSwitch(true);
-                reminder.setSunRepeat(true);
-                reminder.setMonRepeat(true);
-                reminder.setTueRepeat(true);
-                reminder.setWebRepeat(true);
-                reminder.setThurRepeat(true);
-                reminder.setFriRepeat(true);
-                reminder.setSatRepeat(true);
-
-                showWeek(reminder);
-            }
-        });
     }
 
     public void showWeek(final Config.Reminder reminder) {
